--- src/App.js
@@ -1,33 +1,0 @@
-import React from 'react';
-import { View, StatusBar } from 'react-native';
-
-import { Provider as ThemeProvider } from 'react-native-paper';
-
-import Index from './pages/Index';
-
-import Light from './styles/light';
-import Dark from './styles/dark';
-
-var Theme = Light;
-
-var userTheme = "light";
-
-switch (userTheme) {
-    case "light":
-        Theme = Light;
-        break;
-    case "dark":
-        Theme = Dark;
-        break;
-}
-
-export default function App() {
-    return (
-        <ThemeProvider theme={Theme}>
-            <StatusBar backgroundColor="#c66900" barStyle="light-content" />
-            <View style={{ flex: 1, backgroundColor: Theme.colors.background }}>
-                <Index />
-            </View>
-        </ThemeProvider>
-    );
-}
